\name{fineroot}
\alias{fineroot}
\docType{data}
\title{Root Length Density Data}
\description{Root Length Density Data.}
  
\usage{data(fineroot)}
\format{
  A data frame with 511 records and 5 variables: \code{Plant}, identifier of the apple tree, 1-8;\code{Stock}, root stokcing, one of three different root stocks: Mark, MM106 and M26 ; \code{Spacing}, between-row \eqn{\times} within-row spacings, one of the following two: \eqn{4 \times 2} meters and \eqn{5 \times 3} meters; \code{Zone}, inner or outer; \code{RLD}, root length density. 
}

\source{
	\cite{de Silva, H. N., Hall, A. J., Tustin, D. S. and Gandar, P. W. (1999). Analysis of distribution
of root length density of apple trees on different dwarfing rootstocks. \emph{Annals of 
Botany}, 83: 335-345.}

\cite{ Dunn, P.K. and Smyth, G.K. (2005). Series evaluation of Tweedie exponential dispersionmodels densities. \emph{Statistics and Computing}, 15, 267-280.}

}

\examples{
help("fineroot")
}
\keyword{datasets}

