\name{getF}
\alias{getF}
\alias{plotF}

\title{Utitlity function for additive models}
\description{Utility functions to extract and plot smooth terms from a \code{"cpglmm"} object. These are simply copies from the \code{"amer"} package. Detailed documentation can be found at \code{\link[amer]{getF}} and \code{\link[amer]{plotF}}.
}

