\name{insLoss}
\alias{insLoss}
\docType{data}
\title{Insurance Loss Data}
\description{Data from an insurance loss reserving triangle.}
  
\usage{data(insLoss)}
\format{
  A data frame with 55 records and 3 variables: \code{year}, the year when the accident occurs;\code{lag}, development lag; \code{increLoss}, incremental insurance loss in 1000s. 
}

\source{
	\cite{Peters G. W., Shevchenko P. V. and Wuthrich M. V. (2009). Model Uncertainty in Claims Reserving within Tweedie's Compound Poisson Models. \emph{Astin Bulletin}, 39(1),  1-33.}

}

\examples{
help("insLoss")
}
\keyword{datasets}

