\name{cplm-package}
\alias{cplm-package}
\alias{cplm}
\docType{package}
\title{
  Tweedie compound Poisson linear models 
}
\description{This R package provides various methods for fitting Tweedie compound Poisson linear models. The forms of the models that are handled in the package are generalized linear models, mixed-effect models and their corresponding Bayesian models. We provide profiled likelihood approaches, Laplacian approximation and Monte Carlo EM algorithms for non-Bayesian models, and Markov Chain Monte Carlo methods for Bayesian models. All methods allow the index parameter to be estimated from the data.
   
  More information is available
  on the project web site \url{http://code.google.com/p/cplm/}
} 
\details{
\tabular{ll}{
Package: \tab cplm\cr
Type: \tab Package\cr
Version: \tab 0.4-1\cr
Date: \tab 2011-10-08\cr
License: \tab GPL version 2 or later
}
}
\author{
Wayne Zhang <actuary_zhang@hotmail.com>
}
\references{
\cite{ Dunn, P.K. and Smyth, G.K. (2005). Series evaluation of Tweedie exponential dispersion models densities. \emph{Statistics and Computing}, 15, 267-280.}

\cite{McCulloch, C. E. (1997). Maximum likelihood algorithms for generalized linear mixed
models. \emph{Journal of the American Statistical Association}, 92, 162-170.}

}


\keyword{ package }


