\name{NullNum-class}
\docType{class}
\alias{NullNum-class}
\alias{NullList-class}
\alias{NullFunc-class}

\title{Class "NullNum", "NullList" and "NullFunc" }
\description{	Virtual class for \code{c("NULL", "numeric")} , \code{c("NULL","list"} and  \code{c("NULL","function"}}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with these classes in the signature.
}

\keyword{classes}
